backend/requirements.txt
fastapi==0.115.0
pydantic==2.12.5
mangum==0.21.0
PyJWT==2.11.0
cryptography==46.0.4
python-dotenv==1.2.1
boto3==1.42.44
uvicorn==0.40.0

